requirements.txt
Flask==3.0.3
flask_sqlalchemy==3.1.1
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9
python-dotenv==1.0.1
authlib==1.3.1
requests==2.32.3
